Fully use rewrite
