+++ setup.cfg
@@ -1,2 +1,3 @@
 [flake8]
+exclude = __init__.py
 max-line-length = 90
